Adding auto-generated snippets.
