+++ samples/snippets/instances/list_all.py
@@ -1,0 +1,61 @@
+#  Copyright 2022 Google LLC
+#
+#  Licensed under the Apache License, Version 2.0 (the "License");
+#  you may not use this file except in compliance with the License.
+#  You may obtain a copy of the License at
+#
+#      http://www.apache.org/licenses/LICENSE-2.0
+#
+#  Unless required by applicable law or agreed to in writing, software
+#  distributed under the License is distributed on an "AS IS" BASIS,
+#  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+#  See the License for the specific language governing permissions and
+#  limitations under the License.
+
+
+# This file is automatically generated. Please do not modify it directly.
+# Find the relevant recipe file in the samples/recipes or samples/ingredients
+# directory and apply your changes there.
+
+
+# [START compute_instances_list_all]
+from typing import Dict, Iterable
+
+from google.cloud import compute_v1
+
+
+def list_all_instances(
+    project_id: str,
+) -> Dict[str, Iterable[compute_v1.Instance]]:
+    """
+    Return a dictionary of all instances present in a project, grouped by their zone.
+
+    Args:
+        project_id: project ID or project number of the Cloud project you want to use.
+    Returns:
+        A dictionary with zone names as keys (in form of "zones/{zone_name}") and
+        iterable collections of Instance objects as values.
+    """
+    instance_client = compute_v1.InstancesClient()
+    # Use the `max_results` parameter to limit the number of results that the API returns per response page.
+    request = compute_v1.AggregatedListInstancesRequest()
+    request.project = project_id
+    request.max_results = 50
+
+    agg_list = instance_client.aggregated_list(request=request)
+
+    all_instances = {}
+    print("Instances found:")
+    # Despite using the `max_results` parameter, you don't need to handle the pagination
+    # yourself. The returned `AggregatedListPager` object handles pagination
+    # automatically, returning separated pages as you iterate over the results.
+    for zone, response in agg_list:
+        if response.instances:
+            all_instances[zone] = response.instances
+            print(f" {zone}:")
+            for instance in response.instances:
+                print(f" - {instance.name} ({instance.machine_type})")
+    return all_instances
+
+
+# [END compute_instances_list_all]
