+++ samples/snippets/instances/stop.py
@@ -1,0 +1,45 @@
+#  Copyright 2022 Google LLC
+#
+#  Licensed under the Apache License, Version 2.0 (the "License");
+#  you may not use this file except in compliance with the License.
+#  You may obtain a copy of the License at
+#
+#      http://www.apache.org/licenses/LICENSE-2.0
+#
+#  Unless required by applicable law or agreed to in writing, software
+#  distributed under the License is distributed on an "AS IS" BASIS,
+#  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+#  See the License for the specific language governing permissions and
+#  limitations under the License.
+
+
+# This file is automatically generated. Please do not modify it directly.
+# Find the relevant recipe file in the samples/recipes or samples/ingredients
+# directory and apply your changes there.
+
+
+# [START compute_stop_instance]
+from google.cloud import compute_v1
+
+
+def stop_instance(project_id: str, zone: str, instance_name: str):
+    """
+    Stops a stopped Google Compute Engine instance.
+    Args:
+        project_id: project ID or project number of the Cloud project your instance belongs to.
+        zone: name of the zone your instance belongs to.
+        instance_name: name of the instance your want to stop.
+    """
+    instance_client = compute_v1.InstancesClient()
+    op_client = compute_v1.ZoneOperationsClient()
+
+    op = instance_client.stop_unary(
+        project=project_id, zone=zone, instance=instance_name
+    )
+
+    while op.status != compute_v1.Operation.Status.DONE:
+        op = op_client.wait(operation=op.name, zone=zone, project=project_id)
+    return
+
+
+# [END compute_stop_instance]
