+++ samples/snippets/instange_templates/delete.py
@@ -1,0 +1,42 @@
+#  Copyright 2022 Google LLC
+#
+#  Licensed under the Apache License, Version 2.0 (the "License");
+#  you may not use this file except in compliance with the License.
+#  You may obtain a copy of the License at
+#
+#      http://www.apache.org/licenses/LICENSE-2.0
+#
+#  Unless required by applicable law or agreed to in writing, software
+#  distributed under the License is distributed on an "AS IS" BASIS,
+#  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+#  See the License for the specific language governing permissions and
+#  limitations under the License.
+
+
+# This file is automatically generated. Please do not modify it directly.
+# Find the relevant recipe file in the samples/recipes or samples/ingredients
+# directory and apply your changes there.
+
+
+# [START compute_template_delete]
+from google.cloud import compute_v1
+
+
+def delete_instance_template(project_id: str, template_name: str):
+    """
+    Delete an instance template.
+
+    Args:
+        project_id: project ID or project number of the Cloud project you use.
+        template_name: name of the template to delete.
+    """
+    template_client = compute_v1.InstanceTemplatesClient()
+    operation_client = compute_v1.GlobalOperationsClient()
+    op = template_client.delete_unary(
+        project=project_id, instance_template=template_name
+    )
+    operation_client.wait(project=project_id, operation=op.name)
+    return
+
+
+# [END compute_template_delete]
