+++ samples/snippets/instange_templates/get.py
@@ -1,0 +1,43 @@
+#  Copyright 2022 Google LLC
+#
+#  Licensed under the Apache License, Version 2.0 (the "License");
+#  you may not use this file except in compliance with the License.
+#  You may obtain a copy of the License at
+#
+#      http://www.apache.org/licenses/LICENSE-2.0
+#
+#  Unless required by applicable law or agreed to in writing, software
+#  distributed under the License is distributed on an "AS IS" BASIS,
+#  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+#  See the License for the specific language governing permissions and
+#  limitations under the License.
+
+
+# This file is automatically generated. Please do not modify it directly.
+# Find the relevant recipe file in the samples/recipes or samples/ingredients
+# directory and apply your changes there.
+
+
+# [START compute_template_get]
+from google.cloud import compute_v1
+
+
+def get_instance_template(
+    project_id: str, template_name: str
+) -> compute_v1.InstanceTemplate:
+    """
+    Retrieve an instance template, which you can use to create virtual machine
+    (VM) instances and managed instance groups (MIGs).
+
+    Args:
+        project_id: project ID or project number of the Cloud project you use.
+        template_name: name of the template to retrieve.
+
+    Returns:
+        InstanceTemplate object that represents the retrieved template.
+    """
+    template_client = compute_v1.InstanceTemplatesClient()
+    return template_client.get(project=project_id, instance_template=template_name)
+
+
+# [END compute_template_get]
